Extract `OrganisationTableRowPresenter` and `OrganisationLogPresenter` from `OrganisationPresenter`
This gives a clearer separation of our "admin" and "public" concerns
