+++ src/organisations/presenters/organisation.presenter.ts
@@ -1,16 +1,10 @@
 import { I18nService } from 'nestjs-i18n';
 import { Organisation } from './../organisation.entity';
-import { TableRow } from '../../common/interfaces/table-row';
 import { SummaryList } from '../../common/interfaces/summary-list';
-import { escape } from '../../helpers/escape.helper';
 import { formatMultilineString } from '../../helpers/format-multiline-string.helper';
-import { formatDate } from '../../common/utils';
 import { formatLink } from '../../helpers/format-link.helper';
 import { formatEmail } from '../../helpers/format-email.helper';
-import { Profession } from '../../professions/profession.entity';
-import { formatStatus } from '../../helpers/format-status.helper';
 import { formatTelephone } from '../../helpers/format-telephone.helper';
-import { getNationsFromProfessions } from '../../helpers/nations.helper';
 
 interface OrganisationSummaryListOptions {
   classes?: string;
@@ -24,43 +18,6 @@ export class OrganisationPresenter {
     private readonly organisation: Organisation,
     private readonly i18nService: I18nService,
   ) {}
-
-  public tableRow(): TableRow {
-    return [
-      {
-        text: this.organisation.name,
-      },
-      {
-        text: getNationsFromProfessions(this.professions(), this.i18nService),
-      },
-      {
-        text: this.industries(),
-      },
-      {
-        text: this.lastModified,
-      },
-      {
-        text: this.changedBy?.name,
-        attributes: {
-          'data-cy': 'changed-by-user',
-        },
-      },
-      {
-        html: formatStatus(this.organisation.status, this.i18nService),
-      },
-      {
-        html: `<a class="govuk-link" href="/admin/organisations/${
-          this.organisation.id
-        }/versions/${
-          this.organisation.versionId
-        }">${this.i18nService.translate<string>(
-          'organisations.admin.viewDetails',
-          { args: { name: escape(this.organisation.name) } },
-        )}
-        </a>`,
-      },
-    ];
-  }
 
   public summaryList(
     options: OrganisationSummaryListOptions = {},
@@ -159,21 +116,6 @@ export class OrganisationPresenter {
     };
   }
 
-  get changedBy(): { name: string; email: string } {
-    const user = this.organisation.changedByUser;
-
-    return user
-      ? {
-          name: user.name,
-          email: user.email,
-        }
-      : null;
-  }
-
-  get lastModified(): string {
-    return formatDate(this.organisation.lastModified);
-  }
-
   public address(): string {
     return formatMultilineString(this.organisation.address);
   }
@@ -188,26 +130,5 @@ export class OrganisationPresenter {
 
   public url(): string {
     return formatLink(this.organisation.url);
-  }
-
-  private industries(): string {
-    const industries = this.professions()
-      .map((profession) => profession.industries)
-      .flat();
-
-    const industryNames = industries.map((industry) =>
-      this.i18nService.translate<string>(industry.name),
-    );
-
-    return [...new Set(industryNames)].join(', ');
-  }
-
-  private professions(): Profession[] {
-    return this.organisation.professionToOrganisations
-      .filter((relation) => relation.profession)
-      .map((relation) =>
-        Profession.withLatestLiveOrDraftVersion(relation.profession),
-      )
-      .filter((profession) => !!profession);
   }
 }
